Fix up rose querying.

+++ javasrc/edu/brown/cs/fait/query/QueryConstants.java
@@ -81,7 +81,7 @@ interface QueryNode {
    IfaceProgramPoint getProgramPoint();
    IfaceCall getCall();
    
-   void setPriority(int p);
+   void setPriority(double p);
    
 }       // end of interface QueryNode
